+++ Classes/Domain/Model/Registration/Registration.php
@@ -17,6 +17,8 @@ use TYPO3\CMS\Extbase\Persistence\Generic\LazyLoadingProxy;
  */
 class Registration extends AbstractEntity
 {
+    use AttendeesTrait;
+
     /**
      * @var string
      * @Extbase\Validate("StringLength", options={"maximum": 255})
@@ -29,13 +31,6 @@ class Registration extends AbstractEntity
      * @TYPO3\CMS\Extbase\Annotation\ORM\Lazy
      */
     protected $event;
-
-    /**
-     * @var \OliverKlee\FeUserExtraFields\Domain\Model\FrontendUser|null
-     * @phpstan-var FrontendUser|LazyLoadingProxy|null
-     * @TYPO3\CMS\Extbase\Annotation\ORM\Lazy
-     */
-    protected $user;
 
     public function getTitle(): string
     {
@@ -73,23 +68,6 @@ class Registration extends AbstractEntity
         $event = $this->getEvent();
 
         return $event instanceof SingleEvent || $event instanceof EventDate;
-    }
-
-    public function getUser(): ?FrontendUser
-    {
-        $user = $this->user;
-        if ($user instanceof LazyLoadingProxy) {
-            $user = $user->_loadRealInstance();
-            \assert($user instanceof FrontendUser);
-            $this->user = $user;
-        }
-
-        return $user;
-    }
-
-    public function setUser(FrontendUser $user): void
-    {
-        $this->user = $user;
     }
 
     /**